feat(android): create a circleci config (#3)
## Which problem is this PR solving?

Sets up CircleCI integration for the Android SDK.

## Short description of the changes

This is a standard Android CircleCI setup according to
https://circleci.com/developer/orbs/orb/circleci/android

The only change I made was adding a trigger to make it run.

## How to verify that this has the expected result

I don't know how to really test this other than to push it and see that
the project/workflow shows up in CircleCI. Any other suggestions are
appreciated. The `circleci` CLI says the config is "valid":

```
$ circleci config validate .circleci/config.yml
Config file at .circleci/config.yml is valid.
```

I tried to run it locally with the circleci cli, but apparently
"machine" images aren't supported by the cli, and it couldn't download
the image. I tried with a different image, but we'd have to do something
hacky to start the emulator. So it's probably best to follow the
recommended setup here.