feat(android): add basic auto-instrumentation and an example app (#5)
## Which problem is this PR solving?

This adds some of the basic auto-instrumentation from Android OTel to
the Honeycomb SDK.
It also adds an example app that exercises the new auto-instrumentation
and switches the smoke-test to using that app.

## Short description of the changes

1. Adds some of the basic auto-instrumentation from Android OTel we
want. This required updating to the latest OTel version, since the
instrumentation APIs have changed. I manually verified that the
instrumentation works, and tried to document the telemetry it sends in
the README. We will likely tweak this instrumentation or add custom
attributes in the future, but this was enough work for a first pass we
can play with.

2. Moves the smoke-tests into a new "example" app. This was necessary,
because the core library target doesn't have the Activity Lifecycle
events or UI necessary to test the new instrumentation. The intention
was to make a minimal app to exercise the new features. Please note that
a lot of the changes in this PR are boilerplate auto-generated by
Android Studio. In the example app, only the `MainActivity`,
`ExampleApp`, and `HoneycombSmokeTest` classes were significantly
modified.

3. Update the smoke tests to check for the spans from the new features.
A couple of caveats:
a. I am not yet testing the Activity Lifecycle events, because they were
not a goal for this PR, but only incidentally added along with other
instrumentation. We will do another pass on UI instrumentation later, at
which time tests will be added.
b. Crash reporting is not yet part of the smoke-tests. I have manually
tested that it works. But testing crash reporting automatically is
actually really tricky, since the app crashing causes the `androidTest`
harness to consider the tests failed. There are other ways to test it,
but they are complicated, so I'm saving them for later.

4. In the future, we may want to add options for which telemetry is
included or not. But we can figure that out later. So far, these are
pretty basic signals that most every app should want.

5. This also removes the usage of `GlobalOpenTelemetry`. We already
weren't using it, because it's disrecommended for Android, and also
doesn't play well with the testing framework.

## How to verify that this has the expected result

`make smoke`

## Things to consider in review.

* Is the auto-instrumentation being set up correctly? (There is not much
documentation yet.)
* Are the smoke tests sufficient?
* Should we do the example app any differently to make it more useful?
* Is the new documentation in the README clear?


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1207968470870248