feat: add metrics to example app (#11)
## Which problem is this PR solving?
This adds metrics to the example app. It also updates the collector
config to accept metrics as well as updates our smoke tests to assert
that metrics are sent.

It also adds a JsonMetricExporter when `debug` is true. We discovered
that `debug = true` is overwriting existing exporters (ie.
traces/metrics go ONLY to the console when debug is true). That will be
fixed in a subsequent PR.

## How to verify that this has the expected result
Run the sample app, click the "send random metric" button, look in the
collector logs:


![image](https://github.com/user-attachments/assets/f4b15e2b-2c9d-4918-b5db-5ac3263f0de8)

or turn debug on and look in the device logs:


![image](https://github.com/user-attachments/assets/ee960957-b7b2-426e-b507-0a4075071e69)