feat: Implement touch auto-instrumentation on XML-based TextViews (#21)
## Which problem is this PR solving?

- Adds auto-instrumentation for capturing touch events on XML-based
Activities.

## Short description of the changes

The events and attributes are intended to be as close as possible to the
Swift versions, changing attributes only as needed to make sense.

One caveat is that these "clicks" are actually "taps" and don't
necessarily match up to what is a "click" according to a `Button` or
other control itself. This is the best we can do in a reasonable amount
of time, and developers can always use manual instrumentation to capture
the official clicks. We _may_ want to change the behavior in Swift later
to match this more closely, and we _may_ want to rename "click" to
"tap". But those decisions don't need to block this PR.

## How to verify that this has the expected result

`make smoke`