test: add Fragments and more tests to the smoke test app (#31)
## Which problem is this PR solving?

We didn't have any tests that we were properly getting the Activity and
Fragment Lifecycle auto-instrumentations included.

## Short description of the changes

This adds a couple of Fragments to the smoke test app, and an
instrumented test that navigates from the default fragment to a
different one. Then there are new bats assertions to make sure we got
lifecycle events for both fragments.

## How to verify that this has the expected result

The new smoke tests pass.