feat: add manual exception reporting api (#44)
## Which problem is this PR solving?

OpenTelemetry provides auto-instrumentation to automatically log
unhandled exceptions. However, there is no API for logging _handled_
exceptions. This PR adds such an API to the `Honeycomb` class.

## Short description of the changes

The new method is:
```
        fun logException(
            otel: OpenTelemetryRum,
            throwable: Throwable,
            thread: Thread? = null,
        )
```

The logged event conforms with the OpenTelemetry `crash`
auto-instrumentation, which is also consistent with semantic
conventions.

## How to verify that this has the expected result

A new smoke test was added to verify the log is captured.

- [✅] Changelog is updated