fix: fix crash handling and add tests (#87)
## Which problem is this PR solving?

The OpenTelemetry semconv package introduced a breaking change between
1.25.0 and 1.32.0 by removing some symbols. The OpenTelemetry `crash`
auto-instrumentation 0.8.0 was broken by that that change. It is marked
as having 1.25.0 as a dependency. But the Honeycomb SDK depends on
1.32.0. Therefore, gradle resolved the package to 1.32.0 and the `crash`
instrumentation was broken.

The break manifests as uncaught exceptions resulting in the program
hanging indefinitely with an ANR, rather than just crashing.

Our smoke tests didn't catch this breakage, because we weren't testing
the `crash` instrumentation, because of the difficulty of testing an app
crashing.

## Short description of the changes

This change fixes the instrumentation by downgrading the semconv module
back to a known good version. It also adds tests that fail before the
downgrade and pass after. The tests work by calling the uncaught
exception handler directly, rather than trying to cause an actual crash.
That is sufficient to catch failures like the one above.

## How to verify that this has the expected result

The new smoke tests pass.

---

- [X] CHANGELOG is updated
~- [] README is updated with documentation~