feat: add log record processor option (#92)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

Allow us to process log events similarly to how we process spans.

- Closes #<enter issue here>

## Short description of the changes

Adds the option to add a `logRecordProcessor` which gets added to the
logging provider.

## How to verify that this has the expected result
---

- [X] CHANGELOG is updated
- [X] README is updated with documentation