maint: turn off dependabot for otel for now (#93)
## Which problem is this PR solving?

OpenTelemetry Android is still marked as "experimental" and has frequent
changes, and we want our SDK to be a bit more stable. So, for now, turn
off dependabot checks for OTel dependencies. We will update them
manually instead.

## Short description of the changes

Just block all of the OTel dependencies, so that we can only ship
libraries we've tested together.

It's based on the examples here:

https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/controlling-dependencies-updated#ignoring-specific-dependencies

And docs here:

https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference#ignore--

## How to verify that this has the expected result

I'm not really sure how to test it, other than to push it.

---

~- [ ] CHANGELOG is updated~ N/A
~- [ ] README is updated with documentation~ N/A

---------

Co-authored-by: Martin Holman <martin@honeycomb.io>