feat: setup a gradle plugin that generates and injects UUID (#136)
## Short description of the changes
This PR sets up a simple plugin that generates a random UUID and injects
it into the merged android manifest file. It also sets up a properties
file that will contain the generated UUID.
## How to verify that this has the expected result
I added the plugin to the example app. If the plugin fails, then the
build breaks.

---

- [ ] CHANGELOG is updated
- [x] README is updated with documentation