fix: Remove Desugaring for libraries. (#139)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?
Our libraries were compiling with desugaring requiring all installers of
them to need to enable desugaring for their app regardless of SDK
version.

- Closes #<enter issue here>

## Short description of the changes

Removes desugaring requirement in the libraries. Now it only should
require if the SDK version is 26 or below.

## How to verify that this has the expected result

Example app uses desugaring libraries in the smoke test and builds
successfully. Manual testing with desugaring disabled on the example app
was performed as well.

---

- [X] CHANGELOG is updated
- [N/A] README is updated with documentation