fix: downgrade kotlin to RN compatable version (#154)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

React Native is very picky about what version of kotlin it uses and the
current version is 2.0.0 which 2.2.0 is incompatible with.

- Closes #<enter issue here>

## Short description of the changes

Reverts kotlin version and replaces the `putIfAbsent` method that is not
supported in older SDKs.

## How to verify that this has the expected result

Android smoke tests should validate this builds and works as expected.

---

- [X] CHANGELOG is updated
- [n/a] README is updated with documentation