fix: change composable spans to avoid negative times (#207)
## Which problem is this PR solving?

This fixes a subtle bug with Compose render spans sometimes having
really long durations, when the render itself was really fast. The
underlying issue is that we were using `System.currentTimeMillis` to get
the time for the span to end. But the OTel Java SDK uses some trickery
to keep track of timestamps with nano resolution. So, the problem was if
a render finished during the same millisecond that it started, the start
time would have nanos, whereas the end time would have those zeroed out.
When subtracting, the duration of the span would be a small negative
number. When converted to an unsigned uint64_t, the small negative
number became a large positive number.

## Short description of the changes

We can't really introspect the nano-resolution clock of the tracer. So,
instead, I'd like to refactor the spans so that we `end` the span when
it's done, rather than merely recording the time and closing it later.
This should have minimal impact on measurements.

## How to verify that this has the expected result

I don't know a good way to test this.

---

- [X] CHANGELOG is updated
N/A ~- [ ] README is updated with documentation~